todo/src/requirements.txt
colorama==0.4.1
coverage==4.5.4
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.0.2
django-nose==1.4.6
djangorestframework==3.10.2
nose==1.3.7
pinocchio==0.4.2
psycopg2-binary==2.8.3
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
